requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-apscheduler==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
docker==7.1.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==25.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
